requirements.txt
Django==4.2.10
djangorestframework==3.16.0
django-filter==25.1